+++ src/views/toDoModal/toDoModal.vue
@@ -1,5 +1,6 @@
 <template>
-  <div class="modal fade" id="toDoModal" data-backdrop="static" data-keyboard="false" aria-hidden="true">
+  <div class="modal fade" :class="{ 'fullscreen': fullscreenToDoModal }" id="toDoModal" data-backdrop="static"
+    data-keyboard="false" aria-hidden="true">
     <div class="modal-dialog modal-dialog-centered">
       <div class="modal-content">
         <div class="modal-header d-flex">
@@ -8,7 +9,7 @@
               <div v-show="showingCalendar" class="align-items-center date-picker-btn" @click="showCalendar()">
                 <i class="bi-calendar-event mx-2"></i>
                 <datepicker id="todo-date-picker-input" class="py-2" v-model="pickedDate" :locale="language"
-                  :input-format='"dd/MM/yyyy"' />
+                  :input-format='"dd/MM/yyyy"' :weekStartsOn="weekStartOnMonday" />
               </div>
               <div v-show="!showingCalendar" class="align-items-center date-picker-btn">
                 <div class="align-items-center date-picker-btn py-2" id="customListDropDown" data-bs-toggle="dropdown">
@@ -83,8 +84,7 @@
               </li>
             </ul>
             <div>
-              <i class="bi-x close-modal header-menu-icons" data-bs-dismiss="modal"
-                :title="$t('todoDetails.close')"></i>
+              <i class="bi-x close-modal header-menu-icons" data-bs-dismiss="modal" :title="$t('todoDetails.close')"></i>
             </div>
           </div>
         </div>
@@ -97,8 +97,8 @@
                 :class="{ 'completed-task': todo.checked }" @dblclick="editTitle">
                 <span v-html="todoText"></span>
               </label>
-              <label v-show="!editingTitle && todo.text == ''"
-                class="form-check-label todo-title todo-title-empty-title" for="todo-header" @dblclick="editTitle">
+              <label v-show="!editingTitle && todo.text == ''" class="form-check-label todo-title todo-title-empty-title"
+                for="todo-header" @dblclick="editTitle">
                 {{ $t("todoDetails.taskTitle") }}
               </label>
               <input v-show="editingTitle" class="todo-title-input" type="text" v-model="todo.text" ref="titleInput"
@@ -111,9 +111,7 @@
                 <i class="bi-markdown-fill" @mousedown="goToMarkDown" :title="$t('todoDetails.markdown')"></i>
               </div>
               <div v-show="!editingDescription && todo.desc != ''" class="mt-2 todo-description"
-                @dblclick="editDescription">
-                <Markdown :source="todo.desc" />
-              </div>
+                @dblclick="editDescription" v-html="todoDescription"> </div>
               <div v-show="!editingDescription && todo.desc.replace(/^\s*$(?:\r\n?|\n)/gm, '') == ''"
                 @dblclick="editDescription" class="description-empty mt-2">
                 {{ $t("todoDetails.notes") }}
@@ -142,8 +140,8 @@
             <div class="new-sub-task d-flex align-items-center">
               <label for="new-sub-task"><i class="bi-plus-square mx-3"></i></label>
               <input type="text" id="new-sub-task" :placeholder="$t('todoDetails.addSubTask')" autocomplete="off"
-                @blur="addSubTask()" @keyup.enter="addSubTask()" @keyup.esc="cancelAddSubTask()"
-                v-model="newSubTask.text" ref="newSubTask" />
+                @blur="addSubTask()" @keyup.enter="addSubTask()" @keyup.esc="cancelAddSubTask()" v-model="newSubTask.text"
+                ref="newSubTask" />
             </div>
           </ul>
         </div>
@@ -160,13 +158,13 @@
   </div>
 
   <comfirm-modal :id="'removeReModalToDoDetails'" :title="$t('ui.removeRepeatingTask')"
-    :text="$t('ui.repeatingTaskRemoveConfirm')" :ico="'bi-x-circle'" :okText="$t('ui.remove')"
-    @on-ok="removeAllComfirmed" @on-cancel="removeAllCanceled"></comfirm-modal>
+    :text="$t('ui.repeatingTaskRemoveConfirm')" :ico="'bi-x-circle'" :okText="$t('ui.remove')" @on-ok="removeAllComfirmed"
+    @on-cancel="removeAllCanceled"></comfirm-modal>
 </template>
 
 <script>
 import Datepicker from "vue3-datepicker";
-import Markdown from "vue3-markdown-it";
+import MarkdownIt from 'markdown-it';
 import toDoListRepository from "../../repositories/toDoListRepository";
 import moment from "moment";
 import dbRepository from "../../repositories/dbRepository";
@@ -183,6 +181,7 @@ import comfirmModal from "../../components/comfirmModal.vue";
 import linkifyStr from 'linkify-string';
 import ClickHandler from "@manuelernestog/click-handler";
 import tasksHelper from "../../helpers/tasksHelper";
+import markdownTargetBlankLinks from '../../helpers/markdownTargetBlankLinks';
 
 export default {
   name: "toDoModal",
@@ -209,8 +208,12 @@ export default {
       showingCalendar: true,
       loadingView: false,
       options: { target: '_blank', defaultProtocol: 'https' },
-      clickhandler: new ClickHandler()
+      clickhandler: new ClickHandler(),
+      md: new MarkdownIt()
     }
+  },
+  mounted() {
+    markdownTargetBlankLinks.renderBlankLinks(this.md);
   },
   props: {
     selectedTodo: { required: true, type: Object },
@@ -218,7 +221,6 @@ export default {
   components: {
     colorPicker,
     Datepicker,
-    Markdown,
     toastMessage,
     timePicker,
     repeatingEvent,
@@ -372,11 +374,11 @@ export default {
         this.index = this.todoList.length - 1;
         this.todo = this.todoList[this.index];
 
-      if(this.$store.getters.config.autoReorderTasks){
-        this.updateTodoList(newListID,tasksHelper.reorderTasksList( this.todoList));
-      } else {
-        this.updateTodoList(newListID, this.todoList);
-      }
+        if (this.$store.getters.config.autoReorderTasks) {
+          this.updateTodoList(newListID, tasksHelper.reorderTasksList(this.todoList));
+        } else {
+          this.updateTodoList(newListID, this.todoList);
+        }
 
       } else {
         this.loadToDoFormDB(newListID);
@@ -446,7 +448,7 @@ export default {
       };
       this.$store.commit("addTodo", newTodo);
 
-      if(this.$store.getters.config.autoReorderTasks){
+      if (this.$store.getters.config.autoReorderTasks) {
         this.updateTodoList(this.todo.listId, tasksHelper.reorderTasksList(this.$store.getters.todoLists[this.todo.listId]));
       } else {
         this.updateTodoList(this.todo.listId, this.$store.getters.todoLists[this.todo.listId]);
@@ -499,10 +501,10 @@ export default {
       this.updateTodo(false);
     },
     changeSubTaskClickhandler: function (index) {
-      this.clickhandler.handle(function () { this.changeSubTask(index) }.bind(this), function () { this.editSubTask(index) }.bind(this),index);
+      this.clickhandler.handle(function () { this.changeSubTask(index) }.bind(this), function () { this.editSubTask(index) }.bind(this), index);
     },
     changeSubTask: function (index) {
-      if (this.todo.subTaskList[index].checked) {
+      if (this.todo.subTaskList[index].checked && this.moveSubtaskToBotttom) {
         this.todo.subTaskList.push(this.todo.subTaskList.splice(index, 1)[0]);
       }
       this.updateTodo();
@@ -573,6 +575,18 @@ export default {
     },
     todoText: function () {
       return linkifyStr(this.todo.text, this.options);
+    },
+    fullscreenToDoModal: function () {
+      return this.$store.getters.config.fullscreenToDoModal;
+    },
+    moveSubtaskToBotttom: function () {
+      return this.$store.getters.config.moveCompletedSubTaskToBottom;
+    },
+    weekStartOnMonday: function () {
+      return this.$store.getters.config.weekStartOnMonday ? 1 : 0;
+    },
+    todoDescription: function () {
+      return this.md.render(this.todo.desc);
     }
   },
 };
@@ -580,6 +594,34 @@ export default {
 
 <style scoped lang="scss">
 @import "/src/assets/style/globalVars.scss";
+
+
+#toDoModal.fullscreen {
+  .modal-dialog {
+    margin: 0px;
+    height: 80%;
+    width: 80%;
+    max-width: unset;
+    position: absolute;
+    top: 50%;
+    left: 50%;
+    transform: translate(-50%, -50%);
+
+
+    .modal-content {
+      height: 100%;
+      width: 100%;
+    }
+
+    .modal-body {
+      overflow-y: auto;
+    }
+
+    .sub-tasks {
+      max-height: unset;
+    }
+  }
+}
 
 .todo-list-selector .bi-chevron-down {
   @include btn-icon;
@@ -620,7 +662,7 @@ export default {
   border-radius: 3px;
 
   .dark-theme & {
-    border: 2px solid white;
+    border: 1px solid rgba(255, 255, 255, 0.658);
     background-color: unset;
   }
 }
@@ -644,7 +686,7 @@ export default {
   border-radius: 3px;
 
   .dark-theme & {
-    border: 2px solid white;
+    border: 1px solid rgba(255, 255, 255, 0.658);
     color: #c9d1d9;
   }
 }
@@ -790,7 +832,7 @@ export default {
         border-radius: 3px;
 
         .dark-theme & {
-          border: 2px solid white;
+          border: 1px solid rgba(255, 255, 255, 0.658);
           background-color: #21262d;
         }
       }