+++ changelog.md
@@ -1,3 +1,18 @@
+## v2.1.0
+Jun 1, 2023
+
+- Fixing Bugs
+- Improving app performance
+- Refactoring config menu
+- Adding new colors to tasks color picker
+- Upgrading all tech stack libraries
+- Adding new repeating event option for custom weekdays selector
+- Improving splash screen
+- Adding sent subtask to bottom on clomplete setting
+- Todo Description markdown link now redirect to external link in web browser
+- Adding week start on monday setting
+
+
 ## v2.0.0
 Nov 24, 2022